avoid git am *.patch failure when there are no patches
If there are no patches in the current directory to apply git am
will return a non-zero return status which will cause the test to fail.
It should still test for failure, such as when a patch fails to apply,
but it should still work when there are no patches.  Fix this by
checking if there are any patches to apply before running git am.