separate 'build' from 'update'
Separate the 'build' part of the 'update' script so they can be
performed independently.  An update would sync with upstream whereas a
build just builds the kernel and sends a notification.  Now the build
script can be used during a git bisect, which could not use the previous
update script.